:ok_hand: Small fixes on pull request
