+++ src/components/CarouselItem/index.tsx
@@ -1,14 +1,25 @@
 import React, { FC } from "react";
-import { CarouselItemProps, CarouselItemsProps } from "./types";
 import cx from "classnames";
 import styles from "./CarouselItem.module.css";
+import { BaseCarouselItem } from "../../types";
+
+interface CarouselItemProps extends BaseCarouselItem {
+  height: number;
+  width: number;
+}
+
+interface CarouselItemsProps {
+  items: BaseCarouselItem[];
+  height: number;
+  width: number;
+}
 
 const CarouselItem: FC<CarouselItemProps> = ({
   title,
   image,
   width,
   height,
-}) => {
+}: CarouselItemProps) => {
   return (
     <div className={cx(styles.root)}>
       <div className={styles.initialInfoContainer}>
