+++ src/types.ts
@@ -2,3 +2,8 @@ export interface DefaultProps {
   /** Used to add style to components */
   className?: any;
 }
+
+export interface BaseCarouselItem {
+  title: string;
+  image: string;
+}
